requirements.txt
PySide6==6.6.1
reportlab==4.0.9
mariadb
pyyaml
pylint
utils
flake8_docstrings
